requirements.txt
discord.py==2.3.2
flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
aiohttp==3.9.5
Pillow==10.4.0
